Make breadcrumb path the same as stock Magento
The stock Magento vendor/magento/magento-catalog/Block/Breadcrumbs.php does not add the current product page to the end of the breadcrumb list. So if a template assumes the standard Magento breadcrumb code it will insert the current page to the end of the breadcrumbs. The way this extension was previously done would mean that the current product page would show up twice in the breadcrumbs. This modification makes the behavior drop-in compatible with the standard Magento breadcrumb code, so it will output in the same way but still have the full breadcrumb path if the page was found via a product search.